fixed major bug with book
If you ever add_multiplied(0), that's fine and well, but then you would be unable to subtract_multiplied(0) because it would try to divide by 0! I added safety preventing division by 0, but because of that, once the var became 0, it would be forever unable to become anything else.

So, anytime you want to remove a change that multiplied a value by 0, it will simply remove that entry from the book and then go through every remaining entry in the book and do a full_resync() of all of the values. Easy!

What this means is that you should not often multiply vars by 0, or it may begin to impact performance as the var's book grows.