Fixed bug with Bigs in the Book
When adding_change() in the book, if the amount added was a Big, it would store the Big in the book like book[category][source] = amount. So... it saved a reference to that Big! If that Big ever changed somewhere else in the script, it would change inside the book!
Really confusing!
But the fix was to simply change it to book[category][source] = Big.new(amount).