added add_changer()
Anything added in add_changer or the related methods cannot be removed, so anything that is set here should be something that will always have an effect on the variable.

For myself, I am attaching the unit's level as a powerer. It will edit_multiplied based on the effect of the level raised to the power of level. Example: The effect is x2 upon leveling, so then I would add_powerer(2, level). except not a 2, i'd pass in the var that affects the effect. jeez la weez, who even reads this anyway