LoudBoolArray! Default no longer sets current.
Methods should have no side effects! And unfortunately, a very important method like set_default_value() had a big side effect--it also changed the current value! It simply shouldn't do that.

-

LoudBoolArray is a bigger version of LoudBoolPair except that it has limited functionality--its only purpose is to monitor all of the bools you pass into it on new().

I created this to help me keep track of how many available upgrades there were to purchase! An upgrade will always know if it is available for purchase based on the value of this bool array.

-

Added contradict() in LoudBool! It's like copycat, but not like it and would neve be like it.