Big Book changes!
Moved plenty of Book functionality into LoudDict. LoudDict, which kept track of the sum of values in its data dictionary, now PROPERLY keeps track of the sum and is a perfect fit for the needs of Book.

Also, removed the cooldown for Book.changed. When the Book changes, the LoudFloat/LoudInt needs to know immediately. Instead, I put cooldowns on the signals in those classes, so they will not spam your scripts with their internal changes too frequently.