Removed limit in LoudInt. Added LoudIntPair
I removed the filled and emptied signals because the full and empty LoudBools already emit those signals. In my code, I had to replace `my_float_pair.filled` with `my_float_pair.full.became_true`
It's wordier but this is better in the long run because it is less to learn and memorize. We are already familiar with LoudBool functionality, so this is a perfect fit.
The fact that there were two signals to connect to in the class (full.became_true and filled) which were emitted by a single event at the same time (filled 2nd) is just awful